+++ cedar-policy-cli/tests/sample.rs
@@ -30,6 +30,7 @@ use cedar_policy_cli::{
     EvaluateArgs, LinkArgs, PoliciesArgs, PolicyFormat, RequestArgs, ValidateArgs,
 };
 
+use predicates::prelude::*;
 use rstest::rstest;
 
 fn run_check_parse_test(policies_file: impl Into<String>, expected_exit_code: CedarExitCode) {
@@ -998,8 +999,7 @@ fn test_require_policies_for_write() {
 
 #[test]
 fn test_check_parse_json_static_policy() {
-    let json_static_policy: &str =
-        "sample-data/tiny_sandboxes/json-check-parse/static_policy.cedar.json";
+    let json_policy: &str = "sample-data/tiny_sandboxes/json-check-parse/static_policy.cedar.json";
 
     assert_cmd::Command::cargo_bin("cedar")
         .expect("bin exists")
@@ -1007,14 +1007,14 @@ fn test_check_parse_json_static_policy() {
         .arg("--policy-format")
         .arg("json")
         .arg("-p")
-        .arg(json_static_policy)
+        .arg(json_policy)
         .assert()
         .code(0);
 }
 
 #[test]
 fn test_check_parse_json_policy_template() {
-    let json_policy_template: &str =
+    let json_policy: &str =
         "sample-data/tiny_sandboxes/json-check-parse/policy_template.cedar.json";
 
     assert_cmd::Command::cargo_bin("cedar")
@@ -1023,9 +1023,60 @@ fn test_check_parse_json_policy_template() {
         .arg("--policy-format")
         .arg("json")
         .arg("-p")
-        .arg(json_policy_template)
+        .arg(json_policy)
         .assert()
         .code(0);
+}
+
+#[test]
+fn test_check_parse_json_policy_set() {
+    let json_policy: &str = "sample-data/tiny_sandboxes/json-check-parse/policy_set.cedar.json";
+
+    assert_cmd::Command::cargo_bin("cedar")
+        .expect("bin exists")
+        .arg("check-parse")
+        .arg("--policy-format")
+        .arg("json")
+        .arg("-p")
+        .arg(json_policy)
+        .assert()
+        .code(0);
+}
+
+#[test]
+fn test_check_parse_json_policy_mixed_properties() {
+    let json_policy: &str =
+        "sample-data/tiny_sandboxes/json-check-parse/policy_mixed_properties.cedar.json";
+
+    assert_cmd::Command::cargo_bin("cedar")
+        .expect("bin exists")
+        .arg("check-parse")
+        .arg("--policy-format")
+        .arg("json")
+        .arg("-p")
+        .arg(json_policy)
+        .assert()
+        .code(1)
+        .stdout(predicate::str::contains(
+            "matching properties from both formats",
+        ));
+}
+
+#[test]
+fn test_check_parse_json_policy_no_matching_properties() {
+    let json_policy: &str =
+        "sample-data/tiny_sandboxes/json-check-parse/policy_no_matching_properties.cedar.json";
+
+    assert_cmd::Command::cargo_bin("cedar")
+        .expect("bin exists")
+        .arg("check-parse")
+        .arg("--policy-format")
+        .arg("json")
+        .arg("-p")
+        .arg(json_policy)
+        .assert()
+        .code(1)
+        .stdout(predicate::str::contains("no matching properties"));
 }
 
 #[test]